refactor begining processor
