+++ Source/Data processing/mathGenerators.h
@@ -1,14 +1,14 @@
 #include <JuceHeader.h>
-enum FFTOrder
-{
-    order2048 = 11,
-    order4096 = 12,
-    order8192 = 13
-};
 
 template<typename BlockType>
 struct FFTDataGenerator
 {
+    enum FFTOrder
+    {
+        order2048 = 11,
+        order4096 = 12,
+        order8192 = 13
+    };
     /**
      produces the FFT data from an audio buffer.
      */
